+++ sz/src/sz_int8.c
@@ -44,7 +44,7 @@ unsigned int optimize_intervals_int8_1D(int8_t *oriData, size_t dataLength, doub
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -91,7 +91,7 @@ unsigned int optimize_intervals_int8_2D(int8_t *oriData, size_t r1, size_t r2, d
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -145,7 +145,7 @@ unsigned int optimize_intervals_int8_3D(int8_t *oriData, size_t r1, size_t r2, s
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -201,7 +201,7 @@ unsigned int optimize_intervals_int8_4D(int8_t *oriData, size_t r1, size_t r2, s
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
