+++ sz/src/sz_uint16.c
@@ -44,7 +44,7 @@ unsigned int optimize_intervals_uint16_1D(uint16_t *oriData, size_t dataLength, 
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -91,7 +91,7 @@ unsigned int optimize_intervals_uint16_2D(uint16_t *oriData, size_t r1, size_t r
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -145,7 +145,7 @@ unsigned int optimize_intervals_uint16_3D(uint16_t *oriData, size_t r1, size_t r
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -201,7 +201,7 @@ unsigned int optimize_intervals_uint16_4D(uint16_t *oriData, size_t r1, size_t r
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
